requirements.txt
django
django-debug-toolbar
django-extensions
djangorestframework
flake8
ipython
pytest-django
redis
werkzeug